backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flasgger==0.9.7.1
flask-swagger-ui==4.11.1
pyyaml==6.0.1
pyjwt==2.8.0
flask-jwt-extended==4.6.0
cryptography==42.0.5
marshmallow==3.21.1
flask-limiter==3.5.0
alembic==1.13.1
sqlalchemy==2.0.29
flask-socketio==5.3.6
python-socketio==5.11.1
redis==5.0.3
celery==5.3.6
flower==2.0.1

# Python 3.12-compatible ML stack
# PyTorch 2.2.x supports Python 3.12; pair with torchvision 0.17.x
torch==2.2.0
torchvision==0.17.0

# Hugging Face stack - resolve hub/tokenizers/datasets compatibility
transformers==4.35.0
tokenizers==0.14.1
huggingface-hub==0.17.3
datasets==2.14.6
peft==0.7.1
# bitsandbytes provides 8/4-bit quantization (Linux/CUDA environments)
bitsandbytes==0.43.1

# Hyperparameter optimization
optuna==3.6.0

# Imaging for dataset quality (perceptual hashing, previews)
Pillow==10.4.0

# Python 3.12 requires NumPy >=1.26
numpy==1.26.4

# scikit-learn wheels for Python 3.12 are available from 1.4+
scikit-learn==1.4.2

# Ensure build tooling present if sdists appear
setuptools>=68.0.0
wheel>=0.41.0
